refactor: useModal 사용성 개선
