+++ src/hooks/useModal.ts
@@ -1,14 +1,14 @@
 import { useCallback } from 'react';
-import { useRecoilState } from 'recoil';
-import { deleteModalState } from '../service/atom';
+import { RecoilState, useRecoilState } from 'recoil';
+import { ModalType } from '../service/atom';
 import { useBodyScrollLock } from './useBodyScrollLock';
 
 type OpenModalType = {
   callback?: () => void;
 };
 
-export const useModal = () => {
-  const [modalDataState, setModalDataState] = useRecoilState(deleteModalState);
+export const useModal = (modalState: RecoilState<ModalType>) => {
+  const [modalDataState, setModalDataState] = useRecoilState(modalState);
   const { lockScroll, openScroll } = useBodyScrollLock();
 
   const closeModal = useCallback(() => {
